Step 3: Moved NLP and Vision utility scripts to the core package and decoupled from Flask
